smallpangu_v4.0/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "smallpangu"
version = "4.0.0"
description = "轻量级、插件化的个人AI数字分身助手"
readme = "README.md"
authors = [
    {name = "小盘古项目组", email = "smallpangu@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "tomli>=2.0.0; python_version < '3.11'",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]
ai = [
    "openai>=1.0.0",
    "tiktoken>=0.5.0",
    "requests>=2.28.0",
]
ui = [
    "customtkinter>=5.2.0",
    "pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/smallpangu"
Repository = "https://github.com/yourusername/smallpangu.git"
Issues = "https://github.com/yourusername/smallpangu/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.ruff]
line-length = 88
target-version = "py38"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = [
    "E501",  # line too long, handled by black
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
strict = true